+++ pandas/tests/indexes/datetimes/test_scalar_compat.py
@@ -49,9 +49,3 @@ class TestDatetimeIndexOps:
 
         result = getattr(Timestamp(idx[-1]), field)
         assert result == expected
-
-    def test_dti_timestamp_isocalendar_fields(self):
-        idx = tm.makeDateIndex(100)
-        expected = tuple(idx.isocalendar().iloc[-1].to_list())
-        result = idx[-1].isocalendar()
-        assert result == expected
